+++ samplesheets/studyapps/cancer/tests/test_plugins_taskflow.py
@@ -287,6 +287,74 @@ class TestCancerPlugin(
         self.assertEqual(len(sl['data']['bam']['files']), 0)
         self.assertEqual(len(sl['data']['vcf']['files']), 0)
 
+    def test_get_shortcut_links_files_omit(self):
+        """Test get_shortcut_links() with omittable files in iRODS"""
+        self.irods.collections.create(self.source_path)
+        # Create omittable files which come before real ones alphabetically
+        bam_path = os.path.join(
+            self.source_path, '{}_test.bam'.format(SAMPLE_ID_NORMAL)
+        )
+        bam_path_omit = os.path.join(
+            self.source_path, '{}_dragen_evidence.bam'.format(SAMPLE_ID_NORMAL)
+        )
+        vcf_path = os.path.join(
+            self.source_path, '{}_test.vcf.gz'.format(SAMPLE_ID_NORMAL)
+        )
+        vcf_path_omit = os.path.join(
+            self.source_path, '{}_cnv.vcf.gz'.format(SAMPLE_ID_NORMAL)
+        )
+        self.irods.data_objects.create(bam_path)
+        self.irods.data_objects.create(bam_path_omit)
+        self.irods.data_objects.create(vcf_path)
+        self.irods.data_objects.create(vcf_path_omit)
+
+        self.plugin.update_cache(self.cache_name, self.project)
+        study_tables = self.tb.build_study_tables(self.study)
+        sl = self.plugin.get_shortcut_links(
+            self.study, study_tables, case=[SOURCE_ID_NORMAL]
+        )
+
+        self.assertEqual(len(sl['data']['session']['files']), 1)
+        self.assertEqual(len(sl['data']['bam']['files']), 1)
+        self.assertEqual(len(sl['data']['vcf']['files']), 1)
+        self.assertEqual(
+            sl['data']['session']['files'][0]['url'],
+            reverse(
+                'samplesheets.studyapps.cancer:igv',
+                kwargs={'genericmaterial': self.source.sodar_uuid},
+            ),
+        )
+        self.assertEqual(
+            sl['data']['bam']['files'][0]['url'],
+            settings.IRODS_WEBDAV_URL
+            + get_pedigree_file_path('bam', self.source, study_tables),
+        )
+        self.assertEqual(
+            sl['data']['vcf']['files'][0]['url'],
+            settings.IRODS_WEBDAV_URL
+            + get_pedigree_file_path('vcf', self.source, study_tables),
+        )
+
+    def test_get_shortcut_links_files_omit_only(self):
+        """Test get_shortcut_links() with only omittable files in iRODS"""
+        self.irods.collections.create(self.source_path)
+        bam_path_omit = os.path.join(
+            self.source_path, '{}_dragen_evidence.bam'.format(SAMPLE_ID_NORMAL)
+        )
+        vcf_path_omit = os.path.join(
+            self.source_path, '{}_cnv.vcf.gz'.format(SAMPLE_ID_NORMAL)
+        )
+        self.irods.data_objects.create(bam_path_omit)
+        self.irods.data_objects.create(vcf_path_omit)
+        self.plugin.update_cache(self.cache_name, self.project)
+        study_tables = self.tb.build_study_tables(self.study)
+        sl = self.plugin.get_shortcut_links(
+            self.study, study_tables, case=[SOURCE_ID_NORMAL]
+        )
+        self.assertEqual(len(sl['data']['session']['files']), 0)
+        self.assertEqual(len(sl['data']['bam']['files']), 0)
+        self.assertEqual(len(sl['data']['vcf']['files']), 0)
+
     def test_update_cache(self):
         """Test update_cache()"""
         self.plugin.update_cache(self.cache_name, self.project)
@@ -316,5 +384,55 @@ class TestCancerPlugin(
         self.assertEqual(ci['bam'][CASE_IDS[0]], bam_path)
         self.assertEqual(ci['vcf'][CASE_IDS[0]], vcf_path)
         for i in range(1, len(CASE_IDS) - 1):
+            self.assertEqual(ci['bam'][CASE_IDS[i]], None)
+            self.assertEqual(ci['vcf'][CASE_IDS[i]], None)
+
+    def test_update_cache_files_omit(self):
+        """Test update_cache() with omittable files in iRODS"""
+        self.irods.collections.create(self.source_path)
+        # Create omittable files which come before real ones alphabetically
+        bam_path = os.path.join(
+            self.source_path, '{}_test.bam'.format(SAMPLE_ID_NORMAL)
+        )
+        bam_path_omit = os.path.join(
+            self.source_path, '{}_dragen_evidence.bam'.format(SAMPLE_ID_NORMAL)
+        )
+        vcf_path = os.path.join(
+            self.source_path, '{}_test.vcf.gz'.format(SAMPLE_ID_NORMAL)
+        )
+        vcf_path_omit = os.path.join(
+            self.source_path, '{}_cnv.vcf.gz'.format(SAMPLE_ID_NORMAL)
+        )
+        self.irods.data_objects.create(bam_path)
+        self.irods.data_objects.create(bam_path_omit)
+        self.irods.data_objects.create(vcf_path)
+        self.irods.data_objects.create(vcf_path_omit)
+        self.plugin.update_cache(self.cache_name, self.project)
+        ci = self.cache_backend.get_cache_item(
+            APP_NAME, self.cache_name, self.project
+        ).data
+        self.assertEqual(ci['bam'][CASE_IDS[0]], bam_path)
+        self.assertEqual(ci['vcf'][CASE_IDS[0]], vcf_path)
+        for i in range(1, len(CASE_IDS) - 1):
+            self.assertEqual(ci['bam'][CASE_IDS[i]], None)
+            self.assertEqual(ci['vcf'][CASE_IDS[i]], None)
+
+    def test_update_cache_files_omit_only(self):
+        """Test update_cache() with only omittable files in iRODS"""
+        self.irods.collections.create(self.source_path)
+        # Create omittable files which come before real ones alphabetically
+        bam_path_omit = os.path.join(
+            self.source_path, '{}_dragen_evidence.bam'.format(SAMPLE_ID_NORMAL)
+        )
+        vcf_path_omit = os.path.join(
+            self.source_path, '{}_cnv.vcf.gz'.format(SAMPLE_ID_NORMAL)
+        )
+        self.irods.data_objects.create(bam_path_omit)
+        self.irods.data_objects.create(vcf_path_omit)
+        self.plugin.update_cache(self.cache_name, self.project)
+        ci = self.cache_backend.get_cache_item(
+            APP_NAME, self.cache_name, self.project
+        ).data
+        for i in range(0, len(CASE_IDS) - 1):
             self.assertEqual(ci['bam'][CASE_IDS[i]], None)
             self.assertEqual(ci['vcf'][CASE_IDS[i]], None)