+++ samplesheets/studyapps/germline/utils.py
@@ -6,7 +6,7 @@ import logging
 from projectroles.plugins import get_backend_api
 
 from samplesheets.models import GenericMaterial
-from samplesheets.studyapps.utils import FILE_TYPE_SUFFIXES
+from samplesheets.studyapps.utils import check_igv_file_name, FILE_TYPE_SUFFIXES
 from samplesheets.utils import get_index_by_header
 
 
@@ -77,10 +77,13 @@ def get_pedigree_file_path(file_type, source, study_tables):
     if obj_list:
         for query_path in query_paths:
             for obj in obj_list['irods_data']:
-                # NOTE: We no longer expect the SAMPLE name in filenames
-                if obj['path'].startswith(query_path + '/') and obj[
-                    'name'
-                ].lower().endswith(FILE_TYPE_SUFFIXES[file_type]):
+                if (
+                    obj['path'].startswith(query_path + '/')
+                    and obj['name']
+                    .lower()
+                    .endswith(FILE_TYPE_SUFFIXES[file_type])
+                    and check_igv_file_name(obj['name'], file_type)
+                ):
                     file_paths.append(obj['path'])
                     logger.debug('Added path: {}'.format(obj['path']))
     if not file_paths: