+++ CHANGELOG.rst
@@ -23,6 +23,9 @@ Added
     - ``get_latest_file_path()`` helper (#1554)
     - "Not found" element for iRODS modal filter (#1562)
     - Existing iRODS file check in material name editing (#1494)
+    - Omit IGV session files by file name suffix (#1575)
+    - ``SHEETS_IGV_OMIT_BAM`` and ``SHEETS_IGV_OMIT_VCF`` Django settings (#1575)
+    - ``check_igv_file_name()`` in study app utils (#1575)
 
 Changed
 -------